fix: malformed config due to heredoc indents
This commit fixes an issue with Verdaccio where the YAML-formatted
configuration file is malformed due to indent formatting issues. This
would have been fixed by correcting the indents but we have shifted it
out of the Dockerfile so linters can correctly identify the `.yaml` file
format and catch such errors early.